codefresh-package-parser/requirements.txt
requests>=2.28.1
GitPython>=3.1.29
xmltodict>=0.13.0